+++ test/algorithm-segment-tree.cpp
@@ -1,4 +1,5 @@
 // Tests for segment tree.
+#include <rain/algorithm/algorithm.hpp>
 #include <rain/algorithm/segment-tree.hpp>
 #include <rain/literal.hpp>
 #include <rain/time.hpp>
@@ -6,101 +7,95 @@
 #include <cassert>
 #include <iostream>
 
-// Underlying array of LLs, with sum segments.
-class SumTree : public Rain::Algorithm::SegmentTree<long long, long long> {
-	using Rain::Algorithm::SegmentTree<Value, Update>::SegmentTree;
+namespace MaxTreeAux {
+	void aggregateValues(
+		typename std::vector<long long>::reference value,
+		long long const &left,
+		long long const &right) {
+		value = max(left, right);
+	}
+	long long aggregateResults(long long const &left, long long const &right) {
+		return max(left, right);
+	}
+	void apply(
+		typename std::vector<long long>::reference value,
+		long long const &update) {
+		value = max(value, update);
+	}
+}
+using MaxTree = Rain::Algorithm::SegmentTree<
+	long long,
+	long long,
+	long long,
+	-1,
+	MaxTreeAux::aggregateValues,
+	MaxTreeAux::aggregateResults,
+	MaxTreeAux::apply>;
 
-	protected:
-	virtual void aggregate(
-		std::size_t const,
-		typename std::vector<Value>::reference value,
-		Value const &left,
-		Value const &right,
-		std::pair<std::size_t, std::size_t> const &) override {
+namespace SumTreeAux {
+	void aggregateValues(
+		typename std::vector<long long>::reference value,
+		long long const &left,
+		long long const &right) {
 		value = left + right;
 	}
-	virtual Result aggregate(
-		std::size_t const,
-		Result const &left,
-		Result const &right,
-		std::pair<std::size_t, std::size_t> const &) override {
+	long long aggregateResults(long long const &left, long long const &right) {
 		return left + right;
 	}
-	virtual void split(
-		std::size_t const,
-		Update const &update,
-		typename std::vector<Update>::reference left,
-		typename std::vector<Update>::reference right,
-		std::pair<std::size_t, std::size_t> const &) override {
-		left += update;
-		right += update;
+	void apply(
+		typename std::vector<long long>::reference value,
+		long long const &update) {
+		value += update;
 	}
-	virtual Result convert(std::size_t const, Value const &value) override {
-		return value;
-	}
-	virtual void apply(
-		std::size_t const,
-		typename std::vector<Value>::reference value,
-		Update const &update,
-		std::pair<std::size_t, std::size_t> const &range) override {
-		value += update * (range.second - range.first + 1);
-	}
-};
+}
+using SumTree = Rain::Algorithm::SegmentTree<
+	long long,
+	long long,
+	long long,
+	0,
+	SumTreeAux::aggregateValues,
+	SumTreeAux::aggregateResults,
+	SumTreeAux::apply>;
 
 int main() {
 	using namespace Rain::Literal;
 
-	/*
-											0
-					 |                     \
-					 1                      2
-		 |            \            |     \
-		 3             4           5      6
-	|      \      |     \      |   \   | \
-	7       8     9     10    11   12 13 14
-	|  \   | \   | \   | \   | \   |
-	15 16 17 18 19 20 21 22 23 24 25
-	*/
-	SumTree sumTree(11);
-	// [0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0].
-	sumTree.update(0, 5, 7);
-	// [7, 7, 7, 7, 7, 7, 0, 0, 0, 0, 0].
-	assert(sumTree.query(1, 3) == 21);
-	assert(sumTree.query(0, 10) == 42);
-	assert(sumTree.query(1, 1) == 7);
-	assert(sumTree.query(5, 9) == 7);
-	sumTree.update(5, 9, -5);
-	// [7, 7, 7, 7, 7, 2, -5, -5, -5, -5, 0].
-	assert(sumTree.query(5, 5) == 2);
-	assert(sumTree.query(4, 6) == 4);
-	assert(sumTree.query(0, 10) == 17);
-	sumTree.update(0, 10, -1);
-	// [6, 6, 6, 6, 6, 1, -6, -6, -6, -6, -1].
-	sumTree.update(1, 3, 13);
-	// [6, 19, 19, 19, 6, 1, -6, -6, -6, -6, -1].
-	sumTree.update(3, 6, -10);
-	// [6, 19, 19, 9, -4, -9, -16, -6, -6, -6, -1].
-	assert(sumTree.query(0, 7) == 18);
-	assert(sumTree.query(8, 10) == -13);
-	assert(sumTree.query(0, 10) == 5);
-	sumTree.update(0, 3, -5);
-	// [1, 14, 14, 4, -4, -9, -16, -6, -6, -6, -1].
-	sumTree.update(4, 7, 3);
-	// [1, 14, 14, 4, -1, -6, -13, -3, -6, -6, -1].
-	assert(sumTree.query(0, 7) == 10);
-
-	// Time test.
-	auto timeBegin = std::chrono::steady_clock::now();
-	SumTree tree2(100000);
-	for (std::size_t i{0}; i < 100000; i++) {
-		std::pair<std::size_t, std::size_t> range{rand() % 100000, rand() % 100000};
-		if (range.first > range.second) {
-			std::swap(range.first, range.second);
-		}
-		tree2.update(range.first, range.second, rand());
+	{
+		MaxTree tree(10);
+		assert(tree.query(0, 9) == -1);
+		assert(tree.query(1, 1) == -1);
+		assert(tree.query(5, 5) == -1);
+		tree.update(5, 0);
+		assert(tree.query(5, 5) == 0);
+		assert(tree.query(4, 6) == 0);
+		assert(tree.query(1, 1) == -1);
 	}
-	auto timeElapsed = std::chrono::steady_clock::now() - timeBegin;
-	std::cout << "Time elapsed: " << timeElapsed << ".\n";
+
+	{
+		// Same as Fenwick tree tests.
+		SumTree tree(100001);
+		tree.update(0, 100);
+		tree.update(1, 5);
+		tree.update(2, 10);
+		assert(tree.query(0, 2) == 115);
+		tree.update(1, -26);
+		assert(tree.query(0, 9) == 89);
+		// 100, -21, 10.
+
+		tree.update(9, 7);
+		// 100, -21, 10, 0, 0, 0, 0, 0, 0, 7.
+		assert(tree.query(0, 8) == 89);
+		tree.update(2, 5);
+		// 100, -21, 15, 0, 0, 0, 0, 0, 0, 7.
+		tree.update(5, 8);
+		// 100, -21, 15, 0, 0, 8, 0, 0, 0, 7.
+		tree.update(4, -90);
+		// 100, -21, 15, 0, -90, 8, 0, 0, 0, 7.
+		assert(tree.query(4, 5) == -82);
+		assert(tree.query(6, 9) == 7);
+		assert(tree.query(0, 7) == 12);
+		assert(tree.query(0, 100000) == 19);
+	}
 
 	return 0;
 }